chore: int인 항목 변경

+++ src/main/java/upbrella/be/user/service/UserService.java
@@ -5,7 +5,7 @@ import org.springframework.stereotype.Service;
 @Service
 public class UserService {
 
-    public int findUmbrellaBorrowedByUser(long userId) {
+    public long findUmbrellaBorrowedByUser(long userId) {
         return 1;
     }
 }
